requirements.txt
mlflow==1.30.0
pyyaml==6.0
psutil==5.9.2
fastapi==0.104.1
uvicorn==0.23.2
jinja2==3.1.2
python-multipart==0.0.6
pandas==1.4.3
numpy==1.23.2
scikit-learn==1.5.1
pydantic==2.4.2
prometheus-client==0.12.0
starlette-prometheus==0.9.0